Move webpack config merging from index to lib/
