+++ lib/build.js
@@ -6,8 +6,8 @@ const appFromStats = require('./app-from-stats')
 const progress = new ProgressBarPlugin()
 
 module.exports = ({
-  serverConfig,
-  clientConfig
+  serverConfig = {},
+  clientConfig = {}
 }, callback = () => {}) => {
   const serverCompiler = webpack(serverConfig)
