+++ lib/serve.js
@@ -8,9 +8,9 @@ const devServer = require('./dev-server')
 
 module.exports = ({
   port,
-  serverConfig,
-  clientConfig,
-  devConfig,
+  serverConfig = {},
+  clientConfig = {},
+  devConfig = {},
   outputDir
 }) => {
   if (PROD) {
